requirements.txt
asgiref==3.8.1
Django==4.2
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.2
